windows: export getProcessEntry
This function is currently copied in github.com/inconshreveable/mousetrap,
which is widely used in projects (commonly as dependency for spf13/cobra).

Until recently, mousetrap had code for backward-compatibility with old Go
versions, but now that that code was removed, it's a very small wrapper
around getProcessEntry. Exporting this function allows it to be an even
smaller wrapper (most projects would already have golang.org/x/sys as
dependency in some form or another).