unix: add bindings for setattrlist() on macOS
The setattrlist() function can be used to mutate the attributes of files
through a single call. For example, one can perform a chmod(), chown(),
chflags(), [...] all at once.

This change also adds bindings for the UF_* and SF_* flags that are
accepted by chflags(). This makes it possible to use setattrlist() to
perform the equivalent of lchflags(), for which we currently have no
binding.