unix: replace use of strcpy in mkerrors.sh
On OpenBSD-current, clang emits a warning message to standard output
for the use of strcpy, e.g.:

_errors.c(/tmp/_errors-673190.o:(main)): warning: strcpy() is almost always
misused, please use strlcpy()

This message makes it into the Go source being created, causing gofmt
to error on the invalid syntax, and leaving the zerrors file empty.

Using strlcpy would be preferred here, but strncpy is enough to
silence this message, and is more portable.