x/sys/unix: add missing ETHTOOL_FLAG_ constants
Some constants were removed in CL 600516 that included changes for the Linux
kernel 6.10.

This kernel version moved C defines to an enum ethtool_header_flags that was
not picked up by the mkall.sh script.

For enums, there is a perl script that needs to be run manually, and the
output must be added by hand to the list of constants in unix/linux/types.go.

Fixes golang/go#68761